broke up components for easier readability. Utilized useContext for prop drilling
